tests/wpt/web-platform-tests/app-history/navigate/navigate-history-state-replace.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(async () => {
  history.replaceState("state1", "", "#1");
  assert_equals(history.state, "state1");

  let onnavigate_called = false;
  appHistory.onnavigate = () => onnavigate_called = true;
  await appHistory.navigate("#2", { replace: true });
  assert_equals(location.hash, "#2");
  assert_true(onnavigate_called);
  assert_equals(history.state, null);
}, "history.state should be nulled by navigate()");
</script>
